rebase from master and update 10057 base on master
