+++ src/contractTest/java/uk/gov/hmcts/reform/sscs/ContractTestDataProvider.java
@@ -1,5 +1,6 @@
 package uk.gov.hmcts.reform.sscs;
 
+import au.com.dius.pact.consumer.dsl.PactDslJsonBody;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -281,6 +282,151 @@ public class ContractTestDataProvider {
         listUnavailabilityRanges.add(unavailabilityRanges1);
         listUnavailabilityRanges.add(unavailabilityRanges2);
         return listUnavailabilityRanges;
+    }
+
+    protected PactDslJsonBody generateValidHearingGetResponsePactDslJsonBody(LocalDateTime date) {
+        PactDslJsonBody result = new PactDslJsonBody();
+
+        result
+            .object("requestDetails")
+            .stringType("requestTimeStamp", date.toString())
+            .integerType("versionNumber", 123)
+            .stringType("hearingRequestID", "hearingRequestID123")
+            .stringType("status", "status123")
+            .stringType("timeStamp", date.toString())
+            .stringType("hearingGroupRequestId", "hearingGroupRequestId123")
+            .stringType("partiesNotified", date.toString())
+            .closeObject()
+            .object("hearingDetails")
+            .booleanValue("autolistFlag", true)
+            .booleanValue("hearingIsLinkedFlag", true)
+            .booleanValue("privateHearingRequiredFlag", true)
+            .booleanValue("hearingInWelshFlag", true)
+            .stringType("hearingType", "hearingType123")
+            .stringType("leadJudgeContractType", "leadJudgeContractType123")
+            .stringType("listingComments", "listingComments123")
+            .stringType("hearingRequester", "hearingRequester123")
+            .stringType("hearingPriorityType", "hearingPriorityType123")
+            .integerType("numberOfPhysicalAttendees", 123)
+            .integerType("duration", 123)
+            .array("nonStandardHearingDurationReasons")
+            .string("nonStandardHearingDurationReasons1")
+            .string("nonStandardHearingDurationReasons2")
+            .closeArray()
+            .object("hearingWindow")
+            .stringType("dateRangeStart", date.toString())
+            .stringType("dateRangeEnd", date.toString())
+            .stringType("firstDateTimeMustBe", date.toString())
+            .closeObject()
+            .object("panelRequirements")
+            .array("roleType")
+            .string("roleType1")
+            .string("roleType2")
+            .closeArray()
+            .array("authorisationSubType")
+            .string("authorisationSubType1")
+            .string("authorisationSubType2")
+            .closeArray()
+            .array("panelSpecialisms")
+            .string("panelSpecialisms1")
+            .string("panelSpecialisms2")
+            .closeArray()
+            .minArrayLike("panelPreferences", 0, 1)
+            .stringType("memberID", "memberID123")
+            .stringType("memberType", "memberType123")
+            .stringType("requirementType", "requirementType123")
+            .closeObject().closeArray()
+            .closeObject()
+            .minArrayLike("hearingLocations", 0, 1)
+            .stringType("locationType", "locationType123")
+            .stringType("locationId", "locationId123")
+            .closeObject().closeArray()
+            .array("facilitiesRequired")
+            .string("facilitiesRequired1")
+            .closeArray()
+            .closeObject()
+            .object("caseDetails")
+            .booleanValue("caseAdditionalSecurityFlag", true)
+            .booleanValue("caseInterpreterRequiredFlag", true)
+            .booleanValue("caserestrictedFlag", true)
+            .stringType("hmctsServiceCode", "1234")
+            .stringType("caseRef", "1234123412341234")
+            .stringType("externalCaseReference", "externalCaseReference123")
+            .stringType("caseDeepLink", "caseDeepLink123")
+            .stringType("hmctsInternalCaseName", "hmctsInternalCaseName123")
+            .stringType("publicCaseName", "publicCaseName123")
+            .stringType("caseManagementLocationCode", "caseManagementLocationCode123")
+            .stringType("requestTimeStamp", date.toString())
+            .stringType("caseSLAStartDate", date.toString())
+            .minArrayLike("caseCategories", 0, 1)
+            .stringType("categoryType", "categoryType123")
+            .stringType("categoryValue", "categoryValue123")
+            .closeObject().closeArray()
+            .closeObject()
+            .minArrayLike("partyDetails", 0, 1)
+            .stringType("partyID", "partyID123")
+            .stringType("partyType", "partyType123")
+            .stringType("partyRole", "partyRole")
+            .object("individualDetails")
+            .stringType("title", "title123")
+            .stringType("firstName", "firstName123")
+            .stringType("lastName", "lastName123")
+            .stringType("preferredHearingChannel", "preferredHearingChannel123")
+            .stringType("interpreterLanguage", "interpreterLanguage123")
+            .stringType("vulnerabilityDetails", "vulnerabilityDetails123")
+            .stringType("hearingChannelEmail", "hearingChannelEmail123")
+            .stringType("hearingChannelPhone", "07345960795")
+            .booleanValue("vulnerableFlag", true)
+            .array("reasonableAdjustments")
+            .string("reasonableAdjustments1")
+            .closeArray()
+            .minArrayLike("relatedParties", 0, 1)
+            .stringType("relatedPartyID", "relatedPartyID123")
+            .stringType("relationshipType", "relationshipType123")
+            .closeObject().closeArray()
+            .closeObject()
+            .object("organisationDetails")
+            .stringType("name", "name123")
+            .stringType("organisationType", "organisationType123")
+            .stringType("cftOrganisationID", "cftOrganisationID123")
+            .closeObject()
+            .minArrayLike("unavailabilityDOW", 0, 1)
+            .stringType("DOW", "MONDAY")
+            .stringType("DOWUnavailabilityType", "dowUnavailabilityType123")
+            .closeObject().closeArray()
+            .minArrayLike("unavailabilityRanges", 0, 1)
+            .stringType("unavailableFromDate", date.toString())
+            .stringType("unavailableToDate", date.toString())
+            .closeObject().closeArray()
+            .closeObject().closeArray()
+            .object("hearingResponse")
+            .integerType("hearingRequestID", 123)
+            .integerType("versionNumber", 123)
+            .stringType("status", "ACTIVE")
+            .stringType("timeStamp", date.toString())
+            .stringType("listAssistTransactionID", "ListAssistTransactionID123123")
+            .stringType("receivedDateTime", date.toString())
+            .integerType("responseVersion", 321)
+            .stringType("laCaseStatus", "ACTIVE")
+            .stringType("listingStatus", "ACTIVE")
+            .stringType("hearingCancellationReason", "hearingCancellationReason_NO_RESULT")
+            .minArrayLike("hearingDaySchedule", 0, 1)
+            .stringType("listAssistSessionID", "listAssistSessionID123")
+            .stringType("hearingVenueId", "hearingVenueId123")
+            .stringType("hearingRoomId", "hearingRoomId123")
+            .stringType("hearingJudgeId", "hearingJudgeId123")
+            .stringType("panelMemberId", "panelMemberId123")
+            .minArrayLike("attendees", 0, 1)
+            .object("value")
+            .stringType("partyID", "partyID123")
+            .stringType("hearingSubChannel", "hearingSubChannel123")
+            .closeObject()
+            .closeObject().closeArray()
+            .closeObject().closeArray()
+            .closeObject();
+
+
+        return result;
     }
 
 }
