+++ src/main/java/uk/gov/hmcts/reform/sscs/model/ScreenNavigation.java
@@ -1,8 +1,7 @@
-package uk.gov.hmcts.reform.sscs.model.hmc;
+package uk.gov.hmcts.reform.sscs.model;
 
 import lombok.Data;
 
 @Data
 public class ScreenNavigation {
-
 }
