--- src/main/java/uk/gov/hmcts/reform/sscs/model/hmc/Attendees.java
@@ -1,20 +1,0 @@
-package uk.gov.hmcts.reform.sscs.model.hmc;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import uk.gov.hmcts.reform.sscs.exception.ValidationError;
-
-import javax.validation.constraints.Size;
-
-@Data
-@NoArgsConstructor
-@JsonInclude(JsonInclude.Include.NON_NULL)
-public class Attendees {
-    @Size(max = 40, message = ValidationError.PARTY_ID_LENGTH)
-    private String partyID;
-
-    @Size(max = 60, message = ValidationError.HEARING_SUB_CHANNEL_MAX_LENGTH)
-    private String hearingSubChannel;
-
-}
