--- src/main/java/uk/gov/hmcts/reform/sscs/model/hmc/CaseCategory.java
@@ -1,24 +1,0 @@
-package uk.gov.hmcts.reform.sscs.model.hmc;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import uk.gov.hmcts.reform.sscs.exception.ValidationError;
-import uk.gov.hmcts.reform.sscs.validator.EnumPattern;
-
-import javax.validation.constraints.NotEmpty;
-import javax.validation.constraints.Size;
-
-@Data
-@NoArgsConstructor
-@JsonInclude(JsonInclude.Include.NON_NULL)
-public class CaseCategory {
-
-    @NotEmpty(message = ValidationError.CATEGORY_TYPE_EMPTY)
-    @EnumPattern(enumClass = CaseCategoryType.class, fieldName = "categoryType")
-    private String categoryType;
-
-    @NotEmpty(message = ValidationError.CATEGORY_VALUE_EMPTY)
-    @Size(max = 70, message = ValidationError.CATEGORY_VALUE)
-    private String categoryValue;
-}
