--- src/main/java/uk/gov/hmcts/reform/sscs/model/hmc/HearingLocations.java
@@ -1,22 +1,0 @@
-package uk.gov.hmcts.reform.sscs.model.hmc;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import uk.gov.hmcts.reform.sscs.exception.ValidationError;
-import uk.gov.hmcts.reform.sscs.validator.EnumPattern;
-
-import javax.validation.constraints.NotEmpty;
-
-
-@Data
-@NoArgsConstructor
-@JsonInclude(JsonInclude.Include.NON_NULL)
-public class HearingLocations {
-
-    @NotEmpty(message = ValidationError.LOCATION_TYPE_EMPTY)
-    private String locationType;
-
-    @EnumPattern(enumClass = LocationType.class, fieldName = "locationId")
-    private String locationId;
-}
