--- src/main/java/uk/gov/hmcts/reform/sscs/model/hmc/HearingWindow.java
@@ -1,16 +1,0 @@
-package uk.gov.hmcts.reform.sscs.model.hmc;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-
-@Data
-@NoArgsConstructor
-@JsonInclude(JsonInclude.Include.NON_NULL)
-public class HearingWindow {
-
-    private String dateRangeStart;
-    private String dateRangeEnd;
-    private String firstDateTimeMustBe;
-}
