+++ src/main/java/uk/gov/hmcts/reform/sscs/model/single/hearing/HearingResponse.java
@@ -4,8 +4,11 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import uk.gov.hmcts.reform.sscs.ccd.domain.HearingDaySchedule;
+import uk.gov.hmcts.reform.sscs.validator.EnumPattern;
 
 import java.time.LocalDateTime;
+import java.util.List;
 
 
 @Data
@@ -21,4 +24,21 @@ public class HearingResponse {
     private LocalDateTime timeStamp;
 
     private Number versionNumber;
+
+    private String listAssistTransactionID;
+
+    private LocalDateTime receivedDateTime;
+
+    private Integer responseVersion;
+
+    @JsonProperty("laCaseStatus")
+    @EnumPattern(enumClass = ListingCaseStatus.class, fieldName = "listingCaseStatus")
+    private String listingCaseStatus;
+
+    @EnumPattern(enumClass = ListingStatus.class, fieldName = "listingStatus")
+    private String listingStatus;
+
+    private String hearingCancellationReason;
+
+    private List<HearingDaySchedule> hearingDaySchedule;
 }
