+++ src/main/java/uk/gov/hmcts/reform/sscs/model/single/hearing/ListingStatus.java
@@ -1,0 +1,15 @@
+package uk.gov.hmcts.reform.sscs.model.single.hearing;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+@Getter
+public enum ListingStatus {
+    DRAFT("Draft"),
+    PROVISIONAL("Provisional"),
+    FIXED("Fixed");
+
+    private final String listingStatusLabel;
+
+}
